requirements.txt
openai-whisper>=20230314
torch>=2.0.0
sounddevice>=0.4.6
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
chzzkpy>=2.0.0
ollama>=0.1.0
